flask-incidents-azure-storage/requirements.txt
# ===============================================
# Requirements Flask Incidents Azure Storage
# ===============================================

# Flask et extensions
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.53
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
colorama==0.4.6

# Base de données SQL Server / Azure SQL
pyodbc>=5.0.0  # Version 5.2.0 compatible Python 3.12

# Variables d'environnement
python-dotenv==1.0.0

# Azure Storage et authentification
azure-storage-blob==12.17.0
azure-identity==1.15.0
azure-core==1.29.4

# Gestion des fichiers et validation
python-magic-bin==0.4.14  # Pour Windows - détection type MIME
Pillow==10.0.1             # Traitement d'images
# Note: secure_filename est inclus dans Werkzeug (pas de package séparé)

# Optionnel : Azure Key Vault (secrets)
azure-keyvault-secrets==4.7.0

# Développement et debug
Werkzeug==2.3.7

# ===============================================
# Installation recommandée:
# pip install -r requirements.txt
#
# Pour éviter les erreurs de compilation :
# pip install --only-binary=all -r requirements.txt
# ===============================================

# Types de fichiers supportés par défaut :
# - Documents: .pdf, .doc, .docx, .txt, .md
# - Images: .jpg, .jpeg, .png, .gif, .bmp
# - Archives: .zip, .rar, .7z
# - Feuilles de calcul: .xls, .xlsx, .csv
# - Présentations: .ppt, .pptx
#
# Taille maximale par défaut: 16 MB
# Stockage: Azure Blob Storage (conteneur 'incident-documents')